fix(3p/exwm): port fix for exwm#759
exwm#759 is a long-standing EXWM bug around window focus being
incorrectly assigned when switching frames.

There's a mysterious fix in the comments of that issue, which some
people (e.g. ezemtsov) apparently use successfully for a long time.

I'm tired of this bug, and want to try this fix. There's no actual
open PR with this fix, and there's no comments from the maintainer on
this yet, so this commit can be considered experimental.

Change-Id: I9412c004132d0cbe359f987c8c1e0773e625a429