feat(tvix/eval): rewrite Thunk::force() in nonrecursive form
This commit rewrites Thunk::force() so that it is not (directly)
self-recursive.  It maintains a Vec of all the
previously-encountered thunks which point to the one it is currently
forcing, rather than recursively calling itself.

It is still possible for Thunk::force() to recurse, indirectly, by
way of generators::request_enter_lambda().  I have ideas about how
to avoid that.

Change-Id: I6179b8abb2ea0492180fcb347f37595a14665777