feat(tvix/eval): rewrite Thunk::force() in nonrecursive form
This commit rewrites Thunk::force() so that it is not (directly)
self-recursive.  It maintains a Vec of all the
previously-encountered thunks which point to the one it is currently
forcing, rather than recursively calling itself.

It is still possible for Thunk::force() to recurse, indirectly, by
way of generators::request_enter_lambda().  I have ideas about how
to avoid that.

This saves the cost of a round-trip through the generator machinery
for the generators::request_force() which is removed by this commit.

Once a similar transformation has been applied to nix_cmp(),
nix_add(), nix_eq(), and coerce_to_string(), those four functions,
along with Thunk::force(), will make non-tail calls only to each
other.  If they are then merged into a single function
Value::walk(), that function will be tail-recursive, and will not
need to use the generator machinery at all.

The goal here is to use generators for builtins only, so that the
inner "tight loop" of the interpreter doesn't pay the costs that
come with it in terms of performance or complex nonlocal flow
control.

Change-Id: I6179b8abb2ea0492180fcb347f37595a14665777