feat(tvix/castore/tonic): add support for `grpc+inetd:` scheme
Part of b/308. This allows plugging in external store backends without
requiring them to already listen on a tcp port or unix socket, allowing
to "spin up" the process on demand.

In fact, the process isn't even executed at all in case no request to
it is made, similar to how the connection is only established on the
first request (change-able by setting `wait-connect` parameter to `1`).

Some calling conventions still need to be worked out:
 - Should there be a way to configure command arguments to pass to the
   binary from the URL?
 - Should a magic env var be set to enable the mode?
 - Do env vars from the parent process get propagated?

This can be tested as following:

```
❯ cargo build --profile=release-with-debug
   Compiling tvix-castore v0.1.0 (/home/flokli/dev/nixos/code.tvl.fyi/tvix/castore)
   Compiling tvix-store v0.1.0 (/home/flokli/dev/nixos/code.tvl.fyi/tvix/store)
   Compiling tvix-glue v0.1.0 (/home/flokli/dev/nixos/code.tvl.fyi/tvix/glue)
   Compiling tvix-cli v0.1.0 (/home/flokli/dev/nixos/code.tvl.fyi/tvix/cli)
    Finished release-with-debug [optimized + debuginfo] target(s) in 43.29s
❯ BLOB_SERVICE_ADDR=memory:// \
  DIRECTORY_SERVICE_ADDR=memory:// \
  PATH_INFO_SERVICE_ADDR=grpc+inetd:target/release-with-debug/tvix-store \
  target/release-with-debug/tvix-store daemon
  2023-11-15T17:01:39.049086Z  INFO tvix_store: tvix-store listening on [::]:8000
    at store/src/bin/tvix-store.rs:233
```

From another shell:
```
❯ evans -p 8000 -r --host localhost
localhost:8000> package tvix.store.v1

tvix.store.v1@localhost:8000> service PathInfoService

tvix.store.v1.PathInfoService@localhost:8000> call List

[brrr…]
```

As soon as the `List` call is executed (which works), another log line
is printed in the first terminal:

```
  2023-11-15T17:05:09.062945Z  INFO tvix_store: tvix-store listening on inetd
    at store/src/bin/tvix-store.rs:233
```

… confirming a second tvix-store daemon is spun up, which doesn't bind
any ports, but talks to the parent process over stdin and stdout,
and the listing that evans shows us is served through that machinery.

Change-Id: Ie831b98d195f0f76ce75ec0528090f2e1306b9c2