chore(third_party/overlays): disable nix aws integration
The Nix AWS integration is a complete dumpster-fire, the library it
relies on (aws-sdk-cpp) accounts for more build time than cppnix
itself, and its tests fail nondeterministically.  We recently
disabled it for nixpkgs' CI:

  https://github.com/NixOS/nixpkgs/pull/266443

Since it appears TVL does not use the AWS integration (correct me if
I'm wrong) let's disable it here too.  Hopefully someday cppnix will
give up on it and just delete it.

Change-Id: I71668c03379275b2083e59ff7c3a48b3bc7f5e48