feat(nix/buildkite): allow ci.extraSteps to contain derivations
Before this commit, attempting to have an attrvalue of ci.extraSteps
be a derivation would cause failures because normaliseExtraSteps
would choke on the extra derivatrion attributes (system, name,
builder, etc...)

It is desirable to allow CI steps to be derivations, since this
gives us an intuitive way to run them locally with `nix run`.

This CL causes normaliseExtraStep to check if the ci step is a
derivation; if so, the buildkite attributes are taken from
step.passthru.ci instead of directly from step.

Change-Id: I64f50fc966c87f9e330d26ace02eaf73dad45425