feat(tvix/cli): add benchmark for an enormous eval
This comit adds a macrobenchmark benchmark-cross-rocmPackages for
the expression:

  pkgs.pkgsCross.aarch64-multiplatform.rocmPackages_5.hipblas.outPath

This is the largest eval I can think of that doesn't require merging

  https://github.com/NixOS/nixpkgs/pull/269356

beforehand.  This eval is so large that it will wash out any
sampling noise, without having to run multiple iterations (which
would have the cold-cache problem and would artificially exaggerate
startup/shutdown time).

Change-Id: I261d80834db4f4dbac22652d9a16301e4052547f