feat(tvix/eval): nonrecursive coerce_to_string()
After this commit, the only things in value/mod.rs and
value/thunk.rs that use generators are:

  - coerce_to_string() uses generators::request_enter_lambda()
  - Thunk::force() uses generators::request_enter_lambda()

That's it!  There are no other non-builtins callpaths that use the
generator mechanism.  Almost finished.

Technically generators::request_path_import() is also used, but
that's just because the io_handle happens to be part of the VM.
There's no recursion-depth issue there, so the call doesn't need to
go through the generator mechanism (request_path_import() doesn't
call back to the interpreter!)

Change-Id: I83ce5774d49b88fdafdd61160975b4937a435bb0