test(tvix/eval): fix b/343
This commit adds Opcode::OpJumpAndPopIfCatchable, which can be
inserted ahead of most VM operations which expect a boolean on the
stack, in order to handle catchables in branching position properly.

Other than remembering to patch the jump, no other changes should be
required -- this is the reason for the somewhat odd JumpAndPop behavior.

This commit also fixes b/343 by emitting this new opcode when
compiling if-then-else.  There are probably other
yet-to-be-discovered places where we need to do the same thing.

Change-Id: I48de3010014c0bbeba15d34fc0d4800e0bb5a1ef