fix(tvix/eval): resolve incorrect imports when coercing
The default behavior of string coercion in C++ Nix is to weakly coerce
and import to store if necessary. In practice, imports to store never
happen when coercing strongly (although it would be theoretically
possible to achieve this with EvalState::coerceToString) and sometimes
don't when coercing weakly (e.g. for impure builtins, dirOf, …).

To accomodate this behavior we introduce an extra flag for
CoercionKind::Weak and update the call sites to match the C++ Nix
behavior correctly. Note that it is not possible to test this properly
in //tvix/eval tests due to the lack of an appropriate EvalIO
implementation being available. Tests should be added to //tvix/glue
down the line.

Change-Id: I8fb8ab99c7fe08e311d2ba1c36960746bf22f566