refactor(tvix/store/fs): Remove dependence on PathInfoService from TvixStoreFs
To support tvix builds, we want to be able to use the `TvixStoreFs` to
materialize the sandbox's /nix/store filesystem with just the list of
inputs needed for the build. Currently we'd need to implement a new
`PathInfoService` for this which has a larger interface than what is
required by the `TvixStoreFs`. Instead, we'll add a new trait which
includes a minimal, non-Nix-specific, interface for querying for CA
nodes based on file/dir name.

b/341

Change-Id: I928372955017c23b1bf2b37190cbc508a4ed10d5