feat(tvix/eval): track forcing depth within ThunkRepr
This commit adds a two-variant (no "boolean blindness") enum
ForcingDepth to ThunkRepr::Evaluated which tracks how deeply the
thunk has been forced.

This change has two benefits:

1. Thunk::deep_force() no longer needs a ThunkSet -- the "have I
   begun deep-forcing this thunk yet" data which it was used to
   track is now tracked directly in the thunk itself.

2. Calling `deepSeq` on an already-`deepSeq`ed thunk requires no
   additional computation.  This makes it easier to gain performance
   by adding `deepSeq`s to nix code.  It will also assist with
   future strictness-analysis-based optimizations.

Since the ThunkSet struct is now used only for `impl Debug` and for
nothing else, this commit renames it to `ThunkFormatter`.  This
emphasizes that the behavior of its methods do not affect eval.

Change-Id: Ie0207b9fdd7e4565ecd07f6735743652a57a98cf