feat(tvix/eval): avoid allocating a HashSet in Thunk::force()
Previously, we maintained (and allocated) a HashSet just to hold the
thunks that we encountered while forcing a chain-of-thunks.

This commit instead adds an extra field to ThunkRepr::Blackhole, and
forms a linked list of Blackholes as it forces a chain of thunks.
When the final thunk is forced to a value the chain of blackholes is
then walked in reverse order.

With this change, Thunk::force() no longer allocates anything that
can't be placed on the VM's own stack (a HashSet<ThunkRepr> is not a
Value), which will enable using the VM stack to track thunk forcing.
The VM stack has no depth limit, unlike the Rust/CPU stack.

This commit adds a very large comment explaining the situation,
which provides more details than this commit message.

Change-Id: Ife90536942a8243f2c449c170b83eb6e83dbdef0