docs(tvix/eval): document why we thunk ast::Expr::Apply
It initially seemed like not thunking apply nodes was an option.
Although they don't need to have true thunks (i.e. blackholeable)
they do need to be non-strict.  This commit adds the example case
that proves it.

It also notes (and gives the example case) for the third option of
thunking only the *argument* rather than the whole ast::Expr::Apply.
It is unclear if this has any advantage.

Thanks to sterni for having the patience to work through this with
me.

Change-Id: Ia53484ff98d6a8d373901a99a92b81a26306d0d7