chore(tvix): move store/fs to castore/fs
With the recent introduction of the RootNodes trait, there's nothing in
the fs module pulling in tvix-store dependencies, so it can live in
tvix-castore.

tvix-store still has `fuse` and `virtiofs` features, but they now simply
enable these features in the `tvix-castore` crate they depend on.

This allows other crates to make use of it. For example, tvix-build
could it to provide the filesystem available to the builder, which
doesn't need a PathInfoService, but a list of inputs/root nodes.

Change-Id: I989f664827a5a361b23b34368d242d10c157c756