refactor(tvix/store): remove Arc<> from PathInfoService::from_addr
This makes PathInfoService::from_addr return a Box<dyn PathInfoService>,
rather than an Arc<dyn …>, and leaves it up to the consumers to rewrap
it into an Arc where needed.

GRPCPathInfoServiceWrapper embeds its own Arc internally, we definitely
need an Arc for everything gRPC related.

Currently, other places still need it to, but there's ways to get around
that (left for a followup)

Change-Id: Ic83aa2ade6c51912281bd17c7eef7252e152b2d1