fix(tazjin/nixos): disable stub-ld
This thing has been introduced recently and prints annoying warnings
when running manually patchelfed binaries with the default loader
path (which used to work fine!).

This doesn't actually fix running the binaries, which now segfault,
but at least it doesn't print a paragraph of annoying text in the
terminal instead.

Change-Id: Ie0d8c3cc124b71d244e10f3755431fb3401ff81b