feat(tvix/glue): put KnownPaths in an Arc<RwLock<…>>
This ensures it's Send (+Sync).

This was a byproduct while working on cl/10474, and it turned out to be
not necessary, as we could add these builtins late enough to not have to
Send the evaluator around.

It however might still become useful at some point, so let's save this
for later.

Change-Id: Ibe47112240ce469a7c3038c99aa75f7927db3db4